backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
supabase==2.7.4
google-generativeai==0.7.2
httpx==0.27.2
python-dotenv==1.0.1
pytest==8.3.3
pytest-asyncio==0.24.0
gspread==6.1.2
